feat: optimize code structure to facilitate unit testing
